fix UI v2
