+++ quiz-app/src/components/Navbar/index.tsx
@@ -1,113 +1,11 @@
-import {
-  DeliveredProcedureOutlined,
-  FileProtectOutlined,
-  FileUnknownOutlined,
-  HomeOutlined,
-  ProfileOutlined,
-} from '@ant-design/icons';
 import { Layout, Menu } from 'antd';
 import { useAppDispatch } from 'app/hooks';
 import { Contents } from 'components/Contents';
 import Headers from 'components/Header';
+import { defaultMenus, features, settings } from 'constants/menus';
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 const { Sider } = Layout;
-
-interface MenuElement {
-  key: string;
-  to: string;
-  label: string;
-  icon: any;
-  children?: MenuElement[];
-}
-
-const defaultMenus: Array<MenuElement> = [
-  {
-    key: 'home',
-    to: '/',
-    label: 'Home page',
-    icon: <HomeOutlined />,
-    children: [],
-  },
-];
-
-const features: Array<MenuElement> = [
-  {
-    key: 'question-management',
-    to: '/question-management',
-    label: 'Question Management',
-    icon: <FileUnknownOutlined />,
-    children: [],
-  },
-  {
-    key: 'exam-management',
-    to: '/exam-management',
-    label: 'Exam Management',
-    icon: <ProfileOutlined />,
-    children: [],
-  },
-  {
-    key: 'quiz-management',
-    to: '/quiz-management',
-    label: 'Quiz Management',
-    icon: <DeliveredProcedureOutlined />,
-    children: [],
-  },
-];
-
-const settings: Array<MenuElement> = [
-  {
-    key: 'account-management',
-    to: '/account-management',
-    label: 'Account',
-    icon: <ProfileOutlined />,
-    children: [
-      {
-        key: 'change-Password',
-        to: '/change-Password',
-        label: 'Change Password',
-        icon: <HomeOutlined />,
-        children: [],
-      },
-      {
-        key: 'edit-account',
-        to: '/edit-account',
-        label: 'Edit account',
-        icon: <ProfileOutlined />,
-        children: [],
-      },
-    ],
-  },
-  {
-    key: 'users-management',
-    to: '/users-management',
-    label: 'User Management',
-    icon: <FileProtectOutlined />,
-    children: [
-      {
-        key: 'option1',
-        to: '/option1',
-        label: 'option1',
-        icon: <HomeOutlined />,
-        children: [],
-      },
-      {
-        key: 'option2',
-        to: '/option2',
-        label: 'option2',
-        icon: <HomeOutlined />,
-        children: [],
-      },
-      {
-        key: 'option3',
-        to: '/option3',
-        label: 'option3',
-        icon: <ProfileOutlined />,
-        children: [],
-      },
-    ],
-  },
-];
 
 const showSidebarMenu = (token?: string) => {
   let defaults,
@@ -115,7 +13,7 @@ const showSidebarMenu = (token?: string) => {
     itemSettings = [];
   defaults = defaultMenus.map((item: any, index: number) => {
     return (
-      <Menu.Item key={item.key} icon={item.icon}>
+      <Menu.Item key={item.key}>
         <Link to={item.to} className="nav-text">
           {item.label}
         </Link>
@@ -125,7 +23,7 @@ const showSidebarMenu = (token?: string) => {
 
   itemFeatures = features.map((item: any, index: number) => {
     return (
-      <Menu.Item key={item.key} icon={item.icon}>
+      <Menu.Item key={item.key}>
         <Link to={item.to} className="nav-text">
           {item.label}
         </Link>
@@ -135,17 +33,19 @@ const showSidebarMenu = (token?: string) => {
 
   itemSettings = settings.map((item: any) => {
     return item.children.length > 0 ? (
-      <Menu.SubMenu title={item.label} key={item.key} icon={item.icon}>
+      <Menu.SubMenu title={item.label} key={item.key}>
         {item.children.map((subItem: any) => {
           return (
-            <Menu.Item key={subItem.key} icon={subItem.icon}>
-              {subItem.label}
+            <Menu.Item key={subItem.key}>
+              <Link to={subItem.to} className="nav-text">
+                {subItem.label}
+              </Link>
             </Menu.Item>
           );
         })}
       </Menu.SubMenu>
     ) : (
-      <Menu.Item key={item.key} icon={item.icon}>
+      <Menu.Item key={item.key}>
         {item.label}
       </Menu.Item>
     );
