+++ quiz-app/src/constants/configs.ts
