--- quiz-app/src/constants/menus.js
@@ -1,22 +1,0 @@
-export const menus = [
-    {
-      label: 'Home',
-      to: '/',
-      exact: true
-    },
-    {
-      label: 'Category',
-      to: '/category',
-      exact: false
-    },
-    {
-      label: 'Quiz Online',
-      to: '/quiz',
-      exact: false
-    },
-    {
-      label: 'About',
-      to: '/about',
-      exact: false
-    }
-  ]
