+++ quiz-app/src/routers/userRoutes.js
@@ -1,30 +1,30 @@
 import React from 'react';
 
-const NotFound = React.lazy(() => import('components/NotFound'));
-const Logout = React.lazy(() => import('components/Logout'));
-const Home = React.lazy(() => import('features/Home'));
-const QuestionManagement = React.lazy(() => import('features/QuestionManagement'));
-const userRoutes = [
-    {
-        path: '/',
-        main: <Home />
-    },
-    {
-        path: 'home',
-        main: <Home />
-    },
-    {
-        path: 'logout',
-        main: <Logout />
-    },
-    {
-        path: 'question-management',
-        main: <QuestionManagement />
-    },
-    {
-        path: '*',
-        main: <NotFound />
-    }
-]
+export const NotFound = React.lazy(() => import('components/NotFound'));
+export const Logout = React.lazy(() => import('components/Logout'));
+export const Home = React.lazy(() => import('features/Home'));
+export const QuestionManagement = React.lazy(() => import('features/QuestionManagement'));
+export const userRoutes = [
+  {
+    path: '/',
+    main: <Home />,
+  },
+  {
+    path: 'home',
+    main: <Home />,
+  },
+  {
+    path: 'logout',
+    main: <Logout />,
+  },
+  {
+    path: 'question-management',
+    main: <QuestionManagement />,
+  },
+  {
+    path: '*',
+    main: <NotFound />,
+  },
+];
 
-export default userRoutes
+export default userRoutes;
